Add base class for application contexts
